resized the hud tracking elements
git-svn-id: svn://10.0.50.55/puffbomb/trunk@2019 d14347b8-48fd-0310-9e84-97907d434d85

+++ Code/Game/PuffBOMB/GolfGame/GolfGameEngine.cpp
@@ -324,18 +324,18 @@ void cGolfGameEngine::Draw() {
 
 	// Draw player tracking //	
 	for( size_t idx = 0; idx < Player.size(); idx++ ) {
-		ElementTracker( ArrowAnimator, Player[ idx ]->GetRect(), true, -Real( 64 ) );
+		ElementTracker( ArrowAnimator, Player[ idx ]->GetRect(), true, -Real( 48 ) );
 		
-		ElementTracker( PlayerAnimators[ idx ], Player[ idx ]->GetRect(), false, Real( 40 ) );
+		ElementTracker( PlayerAnimators[ idx ], Player[ idx ]->GetRect(), false, Real( 32 ) );
 	}
 	
 	// Draw Points of interest tracking //
 	for ( size_t idx = 0; idx < PointsOfInterest.size(); idx++ ) {
-		ElementTracker( ArrowAnimator, PointsOfInterest[ idx ]->BoundingRect.ToRect(), true, -Real( 64 ) );
+		ElementTracker( ArrowAnimator, PointsOfInterest[ idx ]->BoundingRect.ToRect(), true, -Real( 48 ) );
 		
 		if( PointsOfInterestAnimators.size() == PointsOfInterest.size() )
 		{
-			ElementTracker( PointsOfInterestAnimators[ idx ], PointsOfInterest[ idx ]->BoundingRect.ToRect(), false, Real( 40 ) );
+			ElementTracker( PointsOfInterestAnimators[ idx ], PointsOfInterest[ idx ]->BoundingRect.ToRect(), false, Real( 32 ) );
 		}
 	}
